Flatten LatestUnprocessedVotes into VoteStorage
